Some fixes for fb2013

+++ RimeLib.Content.Frostbite2013_2/Frostbite/Bundles/CasBundle.cs
@@ -135,18 +135,18 @@ public class CasBundle : DbObjectSerializable
         public long Size { get; set; }
         
         [DbObjectField("rangeStart")]
-        public long? RangeStart { get; set; }
+        public uint? RangeStart { get; set; }
         [DbObjectField("rangeEnd")]
-        public long? RangeEnd { get; set; }
+        public uint? RangeEnd { get; set; }
             
         [DbObjectField("logicalOffset")]
-        public long LogicalOffset { get; set; }
+        public uint LogicalOffset { get; set; }
         [DbObjectField("logicalSize")]
-        public long LogicalSize { get; set; }
+        public uint LogicalSize { get; set; }
         
         // For cas deltas
         [DbObjectField("casPatchType")]
-        public long? CasPatchType { get; set; }
+        public long? CasPatchType { get; set; } // TODO: verify type
